- FAT32 초기화 완료 - sac/MassStrage의 일부 사항을 FAT32 작업에 맞게 업데이트 - 일부 파일 tab 수정

+++ yss/inc/mod/qsflash/N25Q128A1.h
@@ -48,8 +48,8 @@ class N25q128a1 : public sac::MassStorage, public sac::QuadspiFlash
 
   public:
 	config::quadspi::Config *getConfig(void);
-	unsigned long getBlockSize(void);
-	unsigned long getNumOfBlock(void);
+	unsigned int getBlockSize(void);
+	unsigned int getNumOfBlock(void);
 	N25q128a1(drv::Quadspi &peri);
 	bool init(void);
 };
